+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+### Unreleased
+
 ### [v0.03](https://github.com/realityforge/galdr/tree/v0.03) (2019-12-29)
 [Full Changelog](https://github.com/realityforge/galdr/compare/v0.02...v0.03)