+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 * Upgrade the `au.com.stocksoftware.idea.codestyle` artifact to version `1.17`.
 * Upgrade the `org.realityforge.braincheck` artifact to version `1.29.0`.
-* Upgrade the `org.realityforge.grim` artifacts to version `0.04`.
+* Upgrade the `org.realityforge.grim` artifacts to version `0.05`.
 * Upgrade the `com.google.truth` artifact to version `0.45`.
 * Upgrade the `com.google.testing.compile` artifact to version `0.18-rf`.
 * Upgrade the `com.squareup` artifact to version `1.13.0`.