+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-### Unreleased
+### [v0.04](https://github.com/realityforge/galdr/tree/v0.04) (2021-03-30) · [Full Changelog](https://github.com/realityforge/galdr/compare/v0.03...v0.04)
 
 * Upgrade the `au.com.stocksoftware.idea.codestyle` artifact to version `1.17`.
 * Upgrade the `org.realityforge.braincheck` artifact to version `1.29.0`.